+++ src/index.ts
@@ -21,3 +21,7 @@ export * from './useSize';
 export * from './useWindowSize';
 export * from './useUrlQueryState';
 export * from './useValueSync';
+export * from './pointUtil';
+export * from './useMousePositionRef';
+export * from './usePan';
+export * from './useScale';