+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - [MINOR] Added `useIsCoreRoutingEnabled` for routing to check if core-routing is used.
 
-## [0.6.3] - 2021-10-29
+### Changed
+- [MAJOR] Updated `Link` component in routing to use href as prop.
 
 ### Added
 - [MINOR] Added `useDeepCompareMemo`