+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `useDeepCompareMemo`
 - [MINOR] Added `useDeepCompareCallback`
 - [MINOR] Added `useIsCoreRoutingEnabled` for routing to check if core-routing is used.
+- [MINOR] Added `usePan` 
+- [MINOR] Added `useMousePositionRef` 
+- [MINOR] Added `useScale` 
+- [MINOR] Added `pointUtil` 
 
 ### Changed
 - [MINOR] Fix `useDeepCompareEffect` to not include the callback as a dependency