+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `useMousePositionRef` 
 - [MINOR] Added `useScale` 
 - [MINOR] Added `pointUtil` 
+- [MINOR] Added default Value  to `useUrlQueryState`
+- [MINOR] Added `useDateUrlQueryState`  
 
 ### Changed
 - [MINOR] Fix `useDeepCompareEffect` to not include the callback as a dependency