+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Added `pointUtil`
 - [MINOR] Added default Value  to `useUrlQueryState`
 - [MINOR] Added `useDateUrlQueryState`
+- [MINOR] Added `MockStorage`
 
 ### Changed
 - [MINOR] Fix `useDeepCompareEffect` to not include the callback as a dependency